+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Some glyphs cause the font renderer to break (https://bugs.mojang.com/browse/MC/issues/MC-13046)
+- Crash with LLibrary
 
 ## [1.3.1] - 2026-3-7